+++ bank-feeds/src/sdk/models/operations/generatecredentials.ts
@@ -52,5 +52,5 @@ export class GenerateCredentialsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }