+++ bank-feeds/src/sdk/models/operations/updatesourceaccount.ts
@@ -58,5 +58,5 @@ export class UpdateSourceAccountResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }